refactoring: rebase FormTextArea fields in postForm component & refactoring usePost hook
