+++ src/features/posts/ui/post/index.ts
@@ -1,2 +1,3 @@
 export * from './post'
+export * from './postForm'
 export * from './usePost'
