+++ src/features/posts/ui/post/post.module.scss
@@ -29,13 +29,6 @@
   padding: 10px;
 }
 
-.char {
-  display: flex;
-  justify-content: end;
-  padding-right: 45px;
-  color: var(--light-900);
-}
-
 .locationContainer {
   display: flex;
   flex-direction: column;
