+++ src/features/posts/ui/post/postForm/index.ts
@@ -1,0 +1,1 @@
+export * from './postForm'
