+++ src/features/posts/ui/post/postForm/postForm.module.scss
@@ -1,0 +1,6 @@
+.char {
+  display: flex;
+  justify-content: end;
+  padding-right: 45px;
+  color: var(--light-900);
+}
